+++ nosql_schema/db/nosqlite/collection_handler.py
@@ -69,6 +69,23 @@ class CollectionHandler(AbstractCollectionHandler):
         query = CollectionHandler.convert_ids(query)
         return self.collection_handle.count(query)
 
+    def create_index(self, keys, **kwargs):
+        """
+        Creates an index on collection
+
+        :param keys: list of keys
+        :param kwargs: further arguments
+        """
+        raise NotImplementedError
+
+    def drop_index(self, name):
+        """
+        Drops an index on collection
+
+        :param name: index name
+        """
+        raise NotImplementedError
+
     @staticmethod
     def convert_ids(query, is_id=False):
         """