+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added `BooleanField`
 - Added `offset` to `find` method
 - Added class `SchemaId` for Id-casting
+- Added `create_index` and `drop_index`
 
 ### Changed
 - Changed `__to_dict` method to public method `to_dict`